wprk with db

+++ server/models/bot_texts.py
@@ -1,4 +1,6 @@
-from sqlalchemy import Column, Integer, String, Text, ForeignKey
+import sqlalchemy
+
+from sqlalchemy import Column, Integer, String, Text, ForeignKey, TIMESTAMP, text, BigInteger
 from sqlalchemy.orm import relationship
 
 from .base import Base
@@ -6,7 +8,7 @@ from .base import Base
 class BotText(Base):
     __tablename__ = 'bot_texts'
 
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     message_key = Column(String(100), nullable=False)
     language_code = Column(String(5), ForeignKey('languages.code'), nullable=False)
     text_content = Column(Text, nullable=False)
